use long for put_pixel and fix byte alignments formula

+++ include/so_long.h
@@ -6,7 +6,7 @@
 /*   By: estettle <estettle@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/15 15:19:22 by estettle          #+#    #+#             */
-/*   Updated: 2024/11/26 21:41:29 by estettle         ###   ########.fr       */
+/*   Updated: 2024/11/27 10:31:41 by estettle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,6 +30,6 @@ typedef struct s_data
 }	t_data;
 
 // Window utils functions
-void	put_pixel(t_data *data, int x, int y, int color);
+void	put_pixel(t_data *data, int x, int y, long color);
 
 #endif //SO_LONG_H
